requirements.txt
librosa
moviepy
pillow
numpy
